App file moved to components
