--- barbershop-landing/src/App.css
@@ -1,5 +1,0 @@
-.login-icon {
-  height: 2rem;
-  cursor: pointer;
-}
-
